# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.BuildingCreateReq:
    properties:
      building_name:
        description: 楼栋名称
        maxLength: 100
        type: string
      building_no:
        description: 楼栋编号
        maxLength: 50
        type: string
      campus_id:
        description: 校区ID
        type: string
      floor_count:
        description: 楼层数
        minimum: 1
        type: integer
      remark:
        description: 备注
        maxLength: 500
        type: string
      room_count:
        description: 房间数
        minimum: 0
        type: integer
    required:
    - building_name
    - building_no
    - campus_id
    - floor_count
    - room_count
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.BuildingListPageReq:
    properties:
      building_name:
        description: 楼栋名称
        type: string
      building_no:
        description: 楼栋编号
        type: string
      campus_id:
        description: 校区ID
        type: string
      currentPage:
        description: 当前页
        type: integer
      pageSize:
        description: 每页条数
        type: integer
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.BuildingResult:
    properties:
      building_name:
        type: string
      building_no:
        type: string
      campus_id:
        type: string
      campus_name:
        description: 校区名称
        type: string
      created_at:
        type: string
      floor_count:
        type: integer
      id:
        type: string
      remark:
        type: string
      room_count:
        type: integer
      room_used_count:
        description: 已分配房间数
        type: integer
      updated_at:
        type: string
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.BuildingUpdateReq:
    properties:
      building_name:
        description: 楼栋名称
        maxLength: 100
        type: string
      building_no:
        description: 楼栋编号
        maxLength: 50
        type: string
      campus_id:
        description: 校区ID
        type: string
      floor_count:
        description: 楼层数
        minimum: 1
        type: integer
      id:
        description: 楼栋ID
        type: string
      remark:
        description: 备注
        maxLength: 500
        type: string
      room_count:
        description: 房间数
        minimum: 0
        type: integer
    required:
    - building_name
    - building_no
    - campus_id
    - floor_count
    - id
    - room_count
    type: object
...
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.PageResult:
    properties:
      currentPage:
        description: 当前页
        type: integer
      list:
        description: 数据列表
      pageSize:
        description: 每页条数
        type: integer
      total:
        description: 总条数
        type: integer
    type: object
...
  github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse:
    properties:
      code:
        example: 400
        type: integer
      err:
        type: object
      msg:
        example: 失败
        type: string
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      data:
        type: object
      msg:
        example: 操作成功
        type: string
    type: object
host: localhost:8880
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for user management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /api/v1/building/byCampus:
    get:
      consumes:
      - application/json
      description: 根据校区ID获取该校区下的所有楼栋信息，用于下拉框选择
      parameters:
      - description: 校区ID
        in: query
        name: campus_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.BuildingResult'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 获取校区下楼栋列表接口
      tags:
      - 楼栋管理
  /api/v1/building/create:
    post:
      consumes:
      - application/json
      description: 创建新的楼栋信息，包含楼栋编号、名称、楼层数、房间数等，编号唯一性自动校验
      parameters:
      - description: 楼栋创建请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.BuildingCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 创建楼栋接口
      tags:
      - 楼栋管理
  /api/v1/building/del:
    post:
      consumes:
      - application/json
      description: 根据楼栋ID列表删除楼栋信息，删除前会检查是否有关联的宿舍
      parameters:
      - description: 删除参数，ids为楼栋ID数组
        in: body
        name: data
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 删除楼栋接口
      tags:
      - 楼栋管理
  /api/v1/building/detail:
    get:
      consumes:
      - application/json
      description: 根据楼栋ID获取楼栋的详细信息，包含所属校区名称和已分配房间数
      parameters:
      - description: 楼栋ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.BuildingResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 获取楼栋详情接口
      tags:
      - 楼栋管理
  /api/v1/building/export:
    get:
      consumes:
      - application/json
      description: 导出楼栋信息为CSV文件，支持按校区和楼栋编号筛选
      parameters:
      - in: query
        name: building_no
        type: string
      - in: query
        name: campus_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: CSV文件
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 批量导出楼栋接口
      tags:
      - 楼栋管理
  /api/v1/building/import:
    post:
      consumes:
      - multipart/form-data
      description: 通过CSV文件批量导入楼栋信息，文件格式：校区名称,楼栋编号,楼栋名称,楼层数,房间数,备注
      parameters:
      - description: CSV文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 批量导入楼栋接口
      tags:
      - 楼栋管理
  /api/v1/building/list:
    get:
      consumes:
      - application/json
      description: 分页查询楼栋信息列表，支持按校区ID、楼栋编号、楼栋名称筛选
      parameters:
      - description: 查询参数
        in: body
        name: data
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.BuildingListPageReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.BuildingResult'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 分页查询楼栋列表接口
      tags:
      - 楼栋管理
  /api/v1/building/update:
    post:
      consumes:
      - application/json
      description: 根据楼栋ID更新楼栋信息，包括楼栋编号、名称、楼层数、房间数等
      parameters:
      - description: 楼栋更新请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.BuildingUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 更新楼栋接口
      tags:
      - 楼栋管理
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"